Small change to get this to work on ios
I tried this on ios, and it didn't work right off. There's some code there that's trying to test whether or not the touch is being used by another FlxAnalog control, and the test isn't right somehow.

The small change I did makes it work (ie the control responds to touches, and a player can be controlled with it), but if the touch floats over to another FlxAnalog, both it and the other FlxAnalog get the touch as input. So not perfect, but I figure you'll know what you were trying to achieve!

I put a demo for FlxAnalog here (https://dl.dropbox.com/u/274922/FlxAnalogDemo.zip
), pretty much copied directly from the Flixel-Android port (http://code.google.com/p/flixel-android/)