Added StartFullscreen parameter to FlxGame (desktop)
Added a new parameter to FlxGame's constructor that allows you to pass
if you want to start the game in Fullscreen or not.
This is useful if you want to say, load the Fullscreen status from a
FlxSave and then decide to start Fullscreen or not based on the user's
options, instead of starting with whatever the Project.xml says, and
then switching later - makes it a little nicer and smoother.