Fix debugger init with cropping camera.
When the camera does not cover the entire game world, the debugger is
initialized with incorrect bounds, so you end up not seeing the debugger
windows. If you resize the window, it's corrected, since FlxDebugger.onResize()
is passed the window dimensions and then corrects it.

This initializes it correctly by just passing the stageWidth / stageHeight.