Fix behavior of fadeIn and fadeOut methods
`fadeIn` and `fadeOut` do not play and stop the music - it is only logical that when the user calls `fadeIn` it is expected for the sound to start playing. 

The same applies to `fadeOut`: when the fading tween completes the sound should call stop instead of continue playing with the volume down.