Flxtext - Prevent blurry lines
Added a method in `FlxText` that insert an extra space to a line that is blurry. This is not a beautful solution, but it works.

Not work if:
If the font have a space character that width is not odd like the default font. (3 pixels)
If `wordWrap` is enabled and is applying a linebreak, because it seems to trim lines.

This solves https://github.com/HaxeFlixel/flixel/issues/1422