make start() public per docs
"FlxTween.PERSIST: stops when it finishes. Unlike ONESHOT, this type of tween stays attached to the core container when it finishes. This means you can keep a reference to this tween and call start() whenever you need it."

From this, I believe it was originally intended that start() be public.

- http://haxeflixel.com/documentation/flxtween/