Make analog sticks emit "fake" digital actions for left/right/up/down movement, and add ID's that make it simple to use them. Makes it easy to use an Analog stick as if it were a DPAD, a very popular input pattern.
This behavior was already implemented on legacy, but the API for it was buried. I added top level FlxGamepadInputID's for the digital events, and refactored things so that Next can emit these events as well.
Also, I added a "value" property to the FlxGamepadButton for storing last frame's analog input value, which should remove the requirement for FlxActions to call check() every frame in order to get accurate JUST_MOVED / JUST_STOPPED behavior.