d'oh (+8 squashed commit)
Squashed commit:

[08aa03d] better "null" check

[dc9a674] fix ifs

[1c9febd] fix crash on legacy

[ffe6838] minor cleanup

[f72cace] this needs to be outside the #if flash conditional, that's probably why I had it at the bottom before

[3ded9b2] style fixes / cleanup

[4d1dd74] quick fix

[7c77870] Make analog sticks emit "fake" digital actions for left/right/up/down movement, and add ID's that make it simple to use them. Makes it easy to use an Analog stick as if it were a DPAD, a very popular input pattern.

This behavior was already implemented on legacy, but the API for it was buried. I added top level FlxGamepadInputID's for the digital events, and refactored things so that Next can emit these events as well.
Also, I added a "value" property to the FlxGamepadButton for storing last frame's analog input value, which should remove the requirement for FlxActions to call check() every frame in order to get accurate JUST_MOVED / JUST_STOPPED behavior. (+6 squashed commit)

Squashed commit:

[1c9febd] fix crash on legacy

[ffe6838] minor cleanup

[f72cace] this needs to be outside the #if flash conditional, that's probably why I had it at the bottom before

[3ded9b2] style fixes / cleanup

[4d1dd74] quick fix

[7c77870] Make analog sticks emit "fake" digital actions for left/right/up/down movement, and add ID's that make it simple to use them. Makes it easy to use an Analog stick as if it were a DPAD, a very popular input pattern.

This behavior was already implemented on legacy, but the API for it was buried. I added top level FlxGamepadInputID's for the digital events, and refactored things so that Next can emit these events as well.
Also, I added a "value" property to the FlxGamepadButton for storing last frame's analog input value, which should remove the requirement for FlxActions to call check() every frame in order to get accurate JUST_MOVED / JUST_STOPPED behavior. (+4 squashed commit)

Squashed commit:

[f72cace] this needs to be outside the #if flash conditional, that's probably why I had it at the bottom before

[3ded9b2] style fixes / cleanup

[4d1dd74] quick fix

[7c77870] Make analog sticks emit "fake" digital actions for left/right/up/down movement, and add ID's that make it simple to use them. Makes it easy to use an Analog stick as if it were a DPAD, a very popular input pattern.

This behavior was already implemented on legacy, but the API for it was buried. I added top level FlxGamepadInputID's for the digital events, and refactored things so that Next can emit these events as well.
Also, I added a "value" property to the FlxGamepadButton for storing last frame's analog input value, which should remove the requirement for FlxActions to call check() every frame in order to get accurate JUST_MOVED / JUST_STOPPED behavior. (+2 squashed commit)

Squashed commit:

[4d1dd74] quick fix

[7c77870] Make analog sticks emit "fake" digital actions for left/right/up/down movement, and add ID's that make it simple to use them. Makes it easy to use an Analog stick as if it were a DPAD, a very popular input pattern.

This behavior was already implemented on legacy, but the API for it was buried. I added top level FlxGamepadInputID's for the digital events, and refactored things so that Next can emit these events as well.
Also, I added a "value" property to the FlxGamepadButton for storing last frame's analog input value, which should remove the requirement for FlxActions to call check() every frame in order to get accurate JUST_MOVED / JUST_STOPPED behavior.