Add completeAll() function to FlxTimerManager
Also, add a note to the docs.  Personally I found reading the FlxTimerManager API doc rather unenlightening and only found out how to use it by grepping the flixel source tree.

The completeAll() function I found useful for my project, where I had several dozen timers going with durations set like "i * 0.05" in a loop.  When I wanted to have it so the user could press a key to skip to the end of the sequence thus created, completeAll() made this a snap.