Change shuffleArray to use unbiased algorithm
For proper shuffling when making card games or anything similar, it's necessary to have an unbiased shuffle algorithm so that every permutation of the array is equally likely. The current shuffleArray function doesn't provide an unbiased result, and depending on the HowManyTimes parameter may do a shuffle of varying quality. This modification replaces the shuffleArray routine with the Fisher-Yates algorithm. It maintains compatibility with current user code by still accepting the HowManyTimes parameter, even though it's not needed anymore.