Add two new signals: debuggerShowed and debuggerHid
Those signals are dispatched when the debugger is showed/hid (usually by pressing the key below ESC). The change was implemented by overriding the setter of the "visible" property in the FlxDebugger class. As a consequence, there is no API changes/breaks.

The new signals are useful for classes interested in monitoring any changes in the debugger visibility; they can now use the (reliable) signals instead of making checks in the update method, for instance.