Make set_visible() work with FLX_NO_DEBUG
The visibilityChanged signal is dispatched only *after* the property "visible" has changed, so signals consumers can use FlxG.debugger.visible to actually check the current visibility of the debugger.

I assume this is the correct implementation of this signal. It is not aligned with the drawDebugChanged signal implementation, which is dispatched *before* the drawDebug property is actually changed.