FlxDefines: fix the Lime version check
Lime 3.0.3's tools still try to build projects with older Lime versions (and are technically able to), so #if (lime >= "3.0.0") never triggers.

However, there's a bug when that older Lime version is also set to legacy like in our case, leading to the lime-legacy.ndll not being copied during the cpp / neko builds and a runtime crash.