FlxSound: fix sounds being stopped too early sometimes
Apparently, we can't rely on the _sound.length values supplied by OpenFL legacy - in Mode, it's 8000 for Mode.ogg (Neko), but in reality it should be ~8698. This led to the sound being stopped too early.

The workaround is to initialize endTime with null instead of length, in which case the sound completes normally.

Issue introduced in f8b43f6.