Update FlxTilemap.hx
made useScaleHack an Int to determine exactly how much scale we want to give it to fix tearing/gaps between tiles.  This was done because sometimes the scale had to be greater than 1 to fill in the gaps between tiles.  This creates a bit more flexibility without disturbing existing behavior.