Make shuffle function slightly more efficient
The shuffle function should have i looping in the range [0,n-2] (where n=array.length). Then we choose a 'j' index to swap with in the range [i,n-1]. However, currently i iterates in the range [0,n-1], which means on the final iteration the function does an unnecessary swap of element (n-1) with itself. I've made a small code change to correct this.