fix FlxAction tests
turns out putting default parameters in activateSet causes weird subtle problems, and we can't put nullables here because this function has to be able to be called every frame (it's how valve recommends you use the steam controller API, among other reasons). Good news is we don't need really need this anymore since we're not generating the default action set from within the action manager when you add actions, so it's okay to require any use of the action set activation API to be 100% explicit.