Add two new properties to customize the flixel.system.Window class
The two new properties added to the class are: _icon and _alwaysOnTop. The _icon property is a reference to the icon being used for the window (null if the window has no icon). Previously, the icon was just added to the window display list, making it difficult to access it for manipulation (particularly from subclasses). The _alwaysOnTop property controls if the window should be put in front of all other windows when clicked. Currently, this is the default behavior, but it prevents more complex customization of the Window class, e.g. tabbed windows.

This change is back-ward compatible.