Color transform fix, closes #2192 (#2195)
The problem was that the shader applies the offsets and multipliers based off of a single flag, hasColorTransform. However, in actually populating the values to be used for these, each colored and hasColorOffsets are checked individually to populate the inputs for the shader.

In the case of a transform with multipliers but no offsets, this was fine, because the lack of offsets had no affect on the result as they effectively default to 0. In the case of offsets but no multipliers, however, treating the multipliers as 0 by default resulted in a black sprite.

I had two options for resolving this. One was to split up the application of offsets and multipliers in the shader and pass in two flag,s one for each, or make it so that if either are true when we are giving the shader the inputs that we populate both inputs instead of only one.

I opted for the latter solution as this means that the shader interface does not change, and I wasn't sure what else might have broken if I did change the shader interface.