Don't set _secondsSinceStart in set_startDelay.
When you set a delay of 3 for the first time before the tween started, _secondsSinceStart becomes 0 * 0 + 3.
So the tween now acts as if 3 seconds have passed since the start. Meaning that it skips the delay of 3 seconds. Furthermore I don't see a good reason as to why you would want to increase the delay after the tween has already started (Started as in had its first update. Essentially every tween has already started, since .start() gets called on creation). And if there was a good reason to do so, I don't see why changing the delay from 3 to 5 seconds should have the _secondsSinceStart be increased by 2 seconds. This might make sense for the loopDelay, but not for the startDelay.