Note that FlxRandom.float() range excludes Max (#2290)
Before this change, the documentation for the float function implied
that it could potentially return any float value within the closed range
[Min, Max]. For example, with the default values of Min = 0 and Max = 1,
both 0 and 1 could be returned.

However, in reality the Max value could never be returned. The
generate() function used internally will return a positive value modulo
the MODULUS constant (i.e. never greater than or equal to MODULUS). This
value is then divided by MODULUS for subsequent calculations, leaving a
value in the range [0, 1) for use. As a result, the Max bound can
mathematically never be reached.

This behavior is canonical and consistent with Haxe's Math.random()
utility. However, the documentation needed to be updated to reflect this
reality.