Added support for Unicode Combining Diacritical Marks in `FlxBitmapText`s For apply Underlines, Umlauts, and other combination characters to your Bitmap Text. More information: https://en.wikipedia.org/wiki/Combining_Diacritical_Marks
Example Usage: add an underline character to your BitmapFont and give it the Id of 818 (Decimal for U+0332), make sure it is as wide as your widest xadvance + letterspacing for your font. Set it's xadvance = 0

In your code, set your `FlxBitmapText.text = "This M\u0332 is Underlined!";` and it should show up with the letter M underlined!