Align FlxText between all targets (#2536)
The desktop and mobile versions of FlxText do not render in the same visual location as web, instead rendering +1 x/y from the origin.

Separately, the code for centering text on a button has always used (x - 1, y +3), even as far back as 2013. Changes to text (possibly through OpenFL) have improved centering and no longer need the (x - 1), which currently causes it to render off-center.